dental-router/requirements.txt
customtkinter>=5.2.0
watchdog>=4.0.0
PyYAML>=6.0.1
